replace old multiselect, new functional, styled for desktop
