+++ client/src/styles/_button.scss
@@ -224,6 +224,10 @@
   color: $justfix-black;
   text-decoration: underline;
   text-underline-position: under;
+  &:focus,
+  &:focus-visible {
+    outline: 0.1rem solid $justfix-grey;
+  }
 }
 
 .buttons.has-addons {
